+++ sites/partners/lib/hooks.ts
@@ -82,56 +82,6 @@ export function useListingsData({ page, limit, userId, sort }: UseListingsDataPr
   }
 }
 
-export function useApplicationsData(
-  pageIndex: number,
-  limit = 10,
-  listingId: string,
-  search: string,
-  orderBy?: EnumApplicationsApiExtraModelOrderBy,
-  orderDir?: EnumApplicationsApiExtraModelOrder
-) {
-  const { applicationsService } = useContext(AuthContext)
-
-  const queryParams = new URLSearchParams()
-  queryParams.append("listingId", listingId)
-  queryParams.append("page", pageIndex.toString())
-  queryParams.append("limit", limit.toString())
-
-  if (search) {
-    queryParams.append("search", search)
-  }
-
-  if (orderBy) {
-    queryParams.append("orderBy", search)
-    queryParams.append("orderDir", orderDir ?? EnumApplicationsApiExtraModelOrder.ASC)
-  }
-
-  const endpoint = `${process.env.backendApiBase}/applications?${queryParams.toString()}`
-
-  const params = {
-    listingId,
-    page: pageIndex,
-    limit,
-  }
-
-  if (search) {
-    Object.assign(params, { search })
-  }
-
-  if (orderBy) {
-    Object.assign(params, { orderBy, orderDir: orderDir ?? "ASC" })
-  }
-
-  const fetcher = () => applicationsService.list(params)
-  const { data, error } = useSWR(endpoint, fetcher)
-
-  return {
-    appsData: data,
-    appsLoading: !error && !data,
-    appsError: error,
-  }
-}
-
 export function useSingleApplicationData(applicationId: string) {
   const { applicationsService } = useContext(AuthContext)
   const backendSingleApplicationsEndpointUrl = `${process.env.backendApiBase}/applications/${applicationId}`
@@ -178,7 +128,57 @@ export function useFlaggedApplicationsList({
     error,
   }
 }
+export function useApplicationsData(
+  currentPage: number,
+  delayedFilterValue: string,
+  limit: number,
+  listingId: string,
+  orderBy?: EnumApplicationsApiExtraModelOrderBy,
+  order?: EnumApplicationsApiExtraModelOrder
+) {
+  const { applicationsService } = useContext(AuthContext)
 
+  const queryParams = new URLSearchParams()
+  queryParams.append("listingId", listingId)
+  queryParams.append("page", currentPage.toString())
+  queryParams.append("limit", limit.toString())
+
+  if (delayedFilterValue) {
+    queryParams.append("search", delayedFilterValue)
+  }
+
+  if (orderBy) {
+    queryParams.append("orderBy", delayedFilterValue)
+    queryParams.append("order", order ?? EnumApplicationsApiExtraModelOrder.ASC)
+  }
+  const endpoint = `${process.env.backendApiBase}/applications?${queryParams.toString()}`
+
+  const params = {
+    listingId,
+    page: currentPage,
+    limit,
+  }
+
+  if (delayedFilterValue) {
+    Object.assign(params, { search: delayedFilterValue })
+  }
+
+  if (orderBy) {
+    Object.assign(params, { orderBy, order: order ?? "ASC" })
+  }
+  const fetcher = () => applicationsService.list(params)
+  const { data, error } = useSWR(endpoint, fetcher)
+
+  const applications = data?.items
+  const appsMeta = data?.meta
+
+  return {
+    applications: applications ?? [],
+    appsMeta,
+    appsLoading: !error && !data,
+    appsError: error,
+  }
+}
 export function useSingleFlaggedApplication(afsId: string) {
   const { applicationFlaggedSetsService } = useContext(AuthContext)
 
@@ -385,7 +385,7 @@ export function useUserList({ page, limit }: UseUserListProps) {
       limit,
       filter: [
         {
-          isPartner: true,
+          isPortalUser: true,
           $comparison: EnumUserFilterParamsComparison["="],
         },
       ],