+++ sites/partners/src/listings/PaperListingForm/UnitForm.tsx
@@ -320,16 +320,32 @@ const UnitForm = ({ onSubmit, onClose, defaultUnit, nextId, draft }: UnitFormPro
   ]
 
   useEffect(() => {
+    if (amiCharts.length === 0 || options.amiCharts.length) return
     setOptions({
+      ...options,
       amiCharts: arrayToFormOptions<AmiChart>(amiCharts, "name", "id"),
+    })
+  }, [amiCharts, options])
+
+  useEffect(() => {
+    if (unitPriorities.length === 0 || options.unitPriorities.length) return
+    setOptions({
+      ...options,
       unitPriorities: arrayToFormOptions<UnitAccessibilityPriorityType>(
         unitPriorities,
         "name",
         "id"
       ),
+    })
+  }, [options, unitPriorities])
+
+  useEffect(() => {
+    if (unitTypes.length === 0 || options.unitTypes.length) return
+    setOptions({
+      ...options,
       unitTypes: arrayToFormOptions<UnitType>(unitTypes, "name", "id", "listings.unit.typeOptions"),
     })
-  }, [amiCharts, unitPriorities, unitTypes])
+  }, [options, unitTypes])
 
   useEffect(() => {
     if (defaultUnit) {
@@ -337,8 +353,7 @@ const UnitForm = ({ onSubmit, onClose, defaultUnit, nextId, draft }: UnitFormPro
       setValue("priorityType.id", defaultUnit.priorityType?.id)
       setValue("unitType.id", defaultUnit.unitType?.id)
     }
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [options])
+  }, [defaultUnit, setValue])
 
   useEffect(() => {
     if (defaultUnit) {